pam-tss-client/requirements.txt
python-tss-sdk>=1.2.2
python-dotenv>=1.0.0
requests>=2.31.0
